+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.0.0 - 2018-12-29]
+### Changed
+- The switch will now always fire the onChange event if the dragStop event occurs <250ms after the dragStart event since that feels like a click.
+- Aria role changed from "checkbox" to "switch".
+
+### Fixed
+- Fix glitch where the clicking the handle wouldn't trigger onChange event when browser window was out of focus on Windows.
+
 ## [3.0.3 - 2018-06-15]
 
 ### Fixed